+++ plugins/variants/src/MultiLinearVariantMatrixDisplay/components/MAFFilterDialog.tsx
@@ -1,3 +1,5 @@
+import React, { useState } from 'react'
+
 import {
   Button,
   Dialog,
@@ -7,7 +9,6 @@ import {
   Typography,
 } from '@mui/material'
 import { observer } from 'mobx-react'
-import React, { useState } from 'react'
 import { makeStyles } from 'tss-react/mui'
 
 const useStyles = makeStyles()({
