+++ plugins/variants/src/MultiLinearVariantMatrixDisplay/configSchema.ts
@@ -2,7 +2,7 @@ import { ConfigurationSchema } from '@jbrowse/core/configuration'
 import { linearBasicDisplayConfigSchemaFactory } from '@jbrowse/plugin-linear-genome-view'
 
 // locals
-import configSchema from '../LinearVariantMatrixRenderer/configSchema'
+import configSchema from '../MultiLinearVariantMatrixRenderer/configSchema'
 
 import type PluginManager from '@jbrowse/core/PluginManager'
