+++ plugins/variants/src/index.ts
@@ -2,9 +2,9 @@ import Plugin from '@jbrowse/core/Plugin'
 
 import ChordVariantDisplayF from './ChordVariantDisplay'
 import LinearVariantDisplayF from './LinearVariantDisplay'
-import LinearVariantMatrixDisplayF from './LinearVariantMatrixDisplay'
-import LinearVariantMatrixRendererF from './LinearVariantMatrixRenderer'
 import MultiLinearVariantDisplayF from './MultiLinearVariantDisplay'
+import LinearVariantMatrixDisplayF from './MultiLinearVariantMatrixDisplay'
+import LinearVariantMatrixRendererF from './MultiLinearVariantMatrixRenderer'
 import MultiVariantRendererF from './MultiVariantRenderer'
 import StructuralVariantChordRendererF from './StructuralVariantChordRenderer'
 import VariantFeatureWidgetF from './VariantFeatureWidget'
